refactor: refactor the directory constructure of the views
