+++ lib/views/home/home_page.dart
@@ -1,7 +1,7 @@
-import 'package:fauth/pages/keys/keys_page.dart';
-import 'package:fauth/pages/settings.dart';
+import 'package:fauth/views/keys/keys_page.dart';
+import 'package:fauth/views/settings/settings_page.dart';
 import 'package:fauth/viewmodels/navigation_viewmodel.dart';
-import 'package:fauth/widgets/adaptive_scaffold.dart';
+import 'package:fauth/views/home/widgets/adaptive_scaffold.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
