+++ lib/views/home/widgets/adaptive_scaffold.dart
