+++ lib/views/keys/widgets/device_info_section.dart
